requirements.txt
backcall==0.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
cycler==0.11.0
decorator==5.1.1
fonttools==4.38.0
idna==3.4
importlib-metadata==6.0.0
ipython==7.34.0
isodate==0.6.1
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
langdetect==1.0.9
MarkupSafe==2.1.2
matplotlib==3.5.3
matplotlib-inline==0.1.6
networkx==2.6.3
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==23.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
rdflib==6.2.0
requests==2.28.2
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
soupsieve==2.3.2.post1
threadpoolctl==3.1.0
torch==1.13.1
torch-geometric==2.2.0
torch-scatter==2.1.0
torch-sparse==0.6.16
tqdm==4.64.1
traitlets==5.9.0
typing_extensions==4.4.0
urllib3==1.26.14
wcwidth==0.2.6
zipp==3.12.0